+++ argilla-server/CHANGELOG.md
@@ -19,6 +19,8 @@ These are the section headers that we use:
 ### Added
 
 - Added new `POST /api/v1/datasets/:dataset_id/export` endpoint to export datasets to Hugging Face Hub using a background job. ([#5709](https://github.com/argilla-io/argilla/pull/5709))
+- Added support to update record fields in `PATCH /api/v1/records/:record_id` endpoint. ([#5685](https://github.com/argilla-io/argilla/pull/5685))
+- Added support to update record fields in `PUT /api/v1/datasets/:dataset_id/records/bulk` endpoint. ([#5685](https://github.com/argilla-io/argilla/pull/5685))
 
 ## [2.5.0](https://github.com/argilla-io/argilla/compare/v2.4.1...v2.5.0)